Move fragments to `ui.v1` package
