+++ src/com/reco1l/osu/multiplayer/RoomScene.kt
@@ -29,7 +29,7 @@ import com.reco1l.osu.multiplayer.Multiplayer.room
 import com.reco1l.osu.ui.MessageDialog
 import com.reco1l.osu.ui.entity.BeatmapButton
 import com.reco1l.osu.ui.entity.ComposedText
-import com.reco1l.osu.ui.SettingsFragment
+import com.osudroid.ui.v1.SettingsFragment
 import com.reco1l.osu.ui.modmenu.ModMenu
 import com.reco1l.osu.updateThread
 import com.reco1l.toolkt.kotlin.runSafe
