+++ src/ru/nsu/ccfit/zuev/osu/game/GameScene.java
@@ -34,7 +34,7 @@ import com.reco1l.osu.hitobjects.FollowPointConnection;
 import com.reco1l.osu.hud.GameplayHUD;
 import com.reco1l.osu.hitobjects.SliderTickSprite;
 import com.reco1l.osu.hud.elements.HUDPPCounter;
-import com.reco1l.osu.ui.BlockAreaFragment;
+import com.osudroid.ui.v1.BlockAreaFragment;
 import com.reco1l.osu.multiplayer.Multiplayer;
 import com.reco1l.osu.multiplayer.RoomScene;
